+++ packages/launchpad/src/lcdapi/auth.ts
@@ -1,6 +1,7 @@
 /* eslint-disable @typescript-eslint/naming-convention */
+import { Pubkey } from "@cosmjs/amino";
+
 import { Coin } from "../coins";
-import { PubKey } from "../types";
 import { LcdClient } from "./lcdclient";
 
 /**
@@ -25,7 +26,7 @@ export interface BaseAccount {
    * [1]: https://github.com/cosmos/cosmos-sdk/pull/5280
    * [2]: https://github.com/cosmos/cosmos-sdk/pull/6749
    */
-  readonly public_key: string | PubKey | null;
+  readonly public_key: string | Pubkey | null;
   /**
    * The account number assigned by the blockchain.
    *